+++ apps/builder/app/builder/builder.tsx
@@ -49,6 +49,12 @@ import { useSearchParams } from "@remix-run/react";
 import { useSyncInitializeOnce } from "~/shared/hook-utils";
 import { BlockingAlerts } from "./features/blocking-alerts";
 import { useStore } from "@nanostores/react";
+import {
+  customComponentMetas,
+  customComponentPropsMetas,
+  registerComponentMetas,
+  registerComponentPropsMetas,
+} from "@webstudio-is/react-sdk";
 
 registerContainers();
 
@@ -239,6 +245,11 @@ export type BuilderProps = {
   authToken?: string;
   authPermit: AuthPermit;
 };
+
+// @todo: Don't do this in builder
+// https://github.com/webstudio-is/webstudio-builder/issues/1545
+registerComponentMetas(customComponentMetas);
+registerComponentPropsMetas(customComponentPropsMetas);
 
 export const Builder = ({
   project,
