+++ packages/react-sdk/src/app/custom-components/index.ts
@@ -1,20 +1,25 @@
+import { imageProps } from "@webstudio-is/image";
+import type { WsComponentPropsMeta } from "../../components/component-meta";
 import { Image } from "./image";
 import { Link } from "./link";
 import { LinkBlock } from "./link-block";
 import { RichTextLink } from "./rich-text-link";
-import { imageProps } from "@webstudio-is/image";
-import type { WsComponentPropsMeta } from "../../components/component-meta";
+import { Form } from "./form";
+import { meta as formMeta, propsMeta as formPropsMeta } from "./form.ws";
 
 export const customComponents = {
   Image,
   Link,
   RichTextLink,
   LinkBlock,
+  Form,
 };
 
 export const customComponentPropsMetas: Record<string, WsComponentPropsMeta> = {
   Image: { props: imageProps },
+  Form: formPropsMeta,
 };
 
-// just for completeness, maybe we add soemthing here later
-export const customComponentMetas = {};
+export const customComponentMetas = {
+  Form: formMeta,
+};
